RESTfullApp HW-4.1 - v2

+++ src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -70,7 +70,7 @@ public class StudentController {
     }
 
     @GetMapping("/getAvgAgeOfStudents")
-    public ResponseEntity<Integer> getAvgAgeOfStudents() {
+    public ResponseEntity<Double> getAvgAgeOfStudents() {
         return ResponseEntity.ok(studentService.getAvgAgeOfStudents());
     }
